+++ src/types/outputs.d.ts
@@ -4,6 +4,7 @@ import {
   InvoiceItem,
   Amount,
   AllocationAllocateeType,
+  ChargeBillableType,
 } from "./api-responses";
 
 export type BookingOutput = {
@@ -194,4 +195,23 @@ export type IssueOutput = {
     name: string | null;
     membership_id: string | null;
   };
+}
+
+export type ChargeOutput = {
+  id: string;
+  description: string;
+  quantity: string;
+  chargeAt: string;
+  carryOver: boolean;
+  accountingCode: string | null;
+  net: string;
+  gross: string;
+  currency: string;
+  costCenterName: string | null;
+  costCenterNumber: string | null;
+  revenueAccountName: string | null;
+  revenueAccountNumber: string | null;
+  spaceId: string;
+  ownerId: string;
+  ownerType: ChargeBillableType;
 };